extract fetching simulator chid processes into a function

+++ webmynd/ios.py
@@ -16,6 +16,21 @@ class IOSRunner(object):
 		if not path.exists(self.sdk):
 			raise ForgeError("Couldn't find the iPhone sdk at: %s" % self.sdk)
 
+	@staticmethod
+	def get_child_processes(target_parent_pid):
+		'Gets processes which have the given pid as their parent'
+		# scrape processes for those with the iphone simulator as the parent
+		list_processes = subprocess.Popen('ps ax -o "pid= ppid="', shell=True, stdout=subprocess.PIPE)
+
+		child_pids = []
+
+		for line in list_processes.stdout:
+			line = line.strip()
+			if line != "":
+				pid, parent_pid = map(int, line.split())
+				if parent_pid == target_parent_pid:
+					child_pids.append(pid)
+
 	def run_iphone_simulator_with(self, file):
 		try:
 			path_to_simulator = path.join(self.sdk, "iPhone Simulator")
@@ -25,23 +40,18 @@ class IOSRunner(object):
 			simulator = subprocess.Popen([path_to_simulator, "-SimulateApplication", path_to_file])
 			LOG.info('simulator pid is %s' % simulator.pid)
 
-			# scrape processes for those with the iphone simulator as the parent
 			# XXX: race condition, the app may not have started yet, so we try a few times.
 			app_pid = None
 			attempts = 0
 
 			while app_pid is None:
 				time.sleep(0.5)
-				list_processes = subprocess.Popen('ps ax -o "pid= ppid="', shell=True, stdout=subprocess.PIPE)
+				child_processes = self.get_child_processes(simulator.pid)
 
-				for line in list_processes.stdout:
-					line = line.strip()
-					if line != "":
-						pid, parent_pid = line.split()
-						if int(parent_pid) == simulator.pid:
-							app_pid = int(pid)
-							LOG.info("pid for iPhone app is: %s" % app_pid)
-							break
+				if child_processes:
+					LOG.info("pid for iPhone app is: %s" % app_pid)
+					app_pid = child_processes[0]
+					break
 
 				if app_pid is None:
 					attempts += 1
@@ -53,5 +63,6 @@ class IOSRunner(object):
 			simulator.communicate()
 
 		finally:
-			LOG.debug('sending kill signal to simulated app...')
-			os.kill(app_pid, signal.SIGTERM)
+			if app_pid is not None:
+				LOG.debug('sending kill signal to simulated app...')
+				os.kill(app_pid, signal.SIGTERM)
